Move database specific code to shared
