+++ androidshared/src/main/java/org/odk/collect/androidshared/sqlite/DatabaseMigrator.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.database;
+package org.odk.collect.androidshared.sqlite;
 
 import android.database.sqlite.SQLiteDatabase;
