+++ collect_app/src/main/java/org/odk/collect/android/database/instances/InstanceDatabaseMigrator.java
@@ -2,7 +2,7 @@ package org.odk.collect.android.database.instances;
 
 import android.database.sqlite.SQLiteDatabase;
 
-import org.odk.collect.android.database.DatabaseMigrator;
+import org.odk.collect.androidshared.sqlite.DatabaseMigrator;
 import org.odk.collect.forms.instances.Instance;
 import org.odk.collect.android.utilities.SQLiteUtils;
